move config to share
